- bugfix by nukehead@gmx.de for temp file deletion upon saving. closes #250, #252, #253, #254, #255
git-svn-id: http://josm.openstreetmap.de/svn/trunk@327 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/actions/SaveActionBase.java
@@ -125,7 +125,7 @@ public abstract class SaveActionBase extends DiskAccessAction {
 					copy(file, tmpFile);
 				}
 				OsmWriter.output(new FileOutputStream(file), new OsmWriter.All(layer.data, false));
-				if (!Main.pref.getBoolean("save.keepbackup"))
+				if (tmpFile != null && !Main.pref.getBoolean("save.keepbackup"))
 					tmpFile.delete();
 			} else if (ExtensionFileFilter.filters[ExtensionFileFilter.CSV].acceptName(file.getPath())) {
 				JOptionPane.showMessageDialog(Main.parent, tr("CSV output not supported yet."));
